Replaced a fatal_error by a handable one.
During the loading of the driver image, the code checks the return code
of mmap.
The original code calls fatal_error followed by a go-to cleanup code if
an error is detected.

Because fatal_error will never return, the go-to instruction is never
executed.
This suggests that the original intention could be
 - fatal_error and nothing else (the go-to instruction should be
deleted)
 - fatal_error is a typo and it should be handable_fatal_error; the
following go-to instruction is correct.

We chose the latter.