Explicit nul-termination of the string returned by readlink.
readlink() writes into a buffer a string that may not be nul-terminated.
In addition, readlink() writes into this buffer up to N bytes then it
is possible to have a buffer with a truncated content (similar to
strncpy)

Checked if a truncation had happen and fail. If not, make sure
that the buffer ends with a 0 byte.